+++ src/component/table/table.tsx
@@ -3,7 +3,7 @@ import { StarRating } from '../star-rating/star-rating'
 import styles from './table.module.scss'
 import { TagList } from '../tag-list/tag-list'
 import { MouseEvent } from 'react'
-import { ProgressBar } from '../progress-bar/progress-bar'
+import { BookProgressBar } from '../progress-bar/book-progress-bar'
 
 interface Props {
     books: Array<Book>
@@ -38,7 +38,7 @@ export function BookTable({ books, onBookClick }: Props) {
                         <td className={styles.pubDate}>{book.metadata.published?.getFullYear() || ''}</td>
                         <td className={styles.pages}>{book.metadata.pages?.toLocaleString() || ''}</td>
                         <td className={styles.duration}>{book.metadata.duration?.toString('verbose') || ''}</td>
-                        <td>{book.progress && <ProgressBar percentage={book.progress} />}</td>
+                        <td>{book.progress && <BookProgressBar book={book} />}</td>
                         <td>
                             <StarRating value={book.metadata.rating || 0} />
                         </td>